# Backend/SuppyLink_User/src/main/java/com/example/suppylink_user/model/User.java
package com.example.suppylink_user.model;


import jakarta.persistence.GeneratedValue;
import jakarta.persistence.GenerationType;
import lombok.Data;
import org.springframework.data.annotation.Id;

import java.util.UUID;

@Data
public class User {
    @Id
    @GeneratedValue(strategy = GenerationType.AUTO)
    private UUID userId;
    private String username;
    private String email;
    private String password;  // Store securely, ideally hashed
}